requirements.txt
pandas>=2.2
openpyxl>=3.1
scikit-learn>=1.5
PyYAML>=6.0
tqdm>=4.66
openai>=1.40
